+++ trunk/academico/desenvolvimento/fonte/shekhinah/src/main/java/br/com/ss/academico/repositorio/SistemaRepositorio.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import br.com.ss.academico.dominio.Sistema;
 
 @Repository
-public interface SistemaRepositorio extends JpaRepository<Sistema, Integer> {
+public interface SistemaRepositorio extends JpaRepository<Sistema, Long> {
 
 	@Query("select r from Sistema r where r.nome like :nome")
 	public List<Sistema> findByNomeLike(@Param("nome") String nome);
